Renamed file from util to utils
